+++ src/core/wallet/wallet.ts
@@ -2,18 +2,17 @@ import { base64, hex } from "@scure/base";
 import * as btc from "@scure/btc-signer";
 import { TAP_LEAF_VERSION, tapLeafHash } from "@scure/btc-signer/payment";
 import { TransactionOutput } from "@scure/btc-signer/psbt";
-
-import { vtxosToTxs } from "../utils/transactionHistory";
-import { BIP21 } from "../utils/bip21";
-import { ArkAddress } from "../core/address";
-import { checkSequenceVerifyScript, VtxoTapscript } from "../core/tapscript";
-import { selectCoins, selectVirtualCoins } from "../utils/coinselect";
-import { getNetwork, Network, NetworkName } from "./networks";
+import { vtxosToTxs } from "../../utils/transactionHistory";
+import { BIP21 } from "../../utils/bip21";
+import { ArkAddress } from "../address";
+import { checkSequenceVerifyScript, VtxoTapscript } from "../tapscript";
+import { selectCoins, selectVirtualCoins } from "../../utils/coinselect";
+import { getNetwork, Network, NetworkName } from "../networks";
 import {
     ESPLORA_URL,
     EsploraProvider,
     OnchainProvider,
-} from "../providers/onchain";
+} from "../../providers/onchain";
 import {
     ArkInfo,
     FinalizationEvent,
@@ -22,147 +21,27 @@ import {
     SigningNoncesGeneratedEvent,
     SigningStartEvent,
     ArkProvider,
-    Output,
-    VtxoInput,
     RestArkProvider,
-} from "../providers/ark";
-import { SignerSession } from "./signingSession";
-import { buildForfeitTx } from "./forfeit";
-import { TxWeightEstimator } from "../utils/txSizeEstimator";
-import { validateConnectorsTree, validateVtxoTree } from "./tree/validation";
-import { Identity } from "./identity";
-
-export interface WalletConfig {
-    network: NetworkName;
-    identity: Identity;
-    esploraUrl?: string;
-    arkServerUrl?: string;
-    arkServerPubKey?: string;
-}
-
-export interface WalletBalance {
-    onchain: {
-        confirmed: number;
-        unconfirmed: number;
-        total: number;
-    };
-    offchain: {
-        swept: number;
-        settled: number;
-        pending: number;
-        total: number;
-    };
-    total: number;
-}
-
-export interface SendBitcoinParams {
-    address: string;
-    amount: number;
-    feeRate?: number;
-    memo?: string;
-}
-
-export interface Recipient {
-    address: string;
-    amount: number;
-}
-
-// SpendableVtxo embed the forfeit script to use as spending path for the boarding utxo or vtxo
-export type SpendableVtxo = VtxoInput & {
-    forfeitScript: string;
-};
-
-export interface SettleParams {
-    inputs: (string | SpendableVtxo)[];
-    outputs: Output[];
-}
-
-// VtxoTaprootAddress embed the tapscripts composing the address
-// it admits the internal key is the unspendable x-only public key
-export interface VtxoTaprootAddress {
-    address: string;
-    scripts: {
-        exit: string[];
-        forfeit: string[];
-    };
-}
-
-export interface AddressInfo {
-    onchain: string;
-    offchain?: VtxoTaprootAddress;
-    boarding?: VtxoTaprootAddress;
-    bip21: string;
-}
-
-export interface TapscriptInfo {
-    offchain?: string[];
-    boarding?: string[];
-}
-
-export interface Status {
-    confirmed: boolean;
-    block_height?: number;
-    block_hash?: string;
-    block_time?: number;
-}
-
-export interface VirtualStatus {
-    state: "pending" | "settled" | "swept" | "spent";
-    batchTxID?: string;
-    batchExpiry?: number;
-}
-
-export interface Outpoint {
-    txid: string;
-    vout: number;
-}
-
-export interface Coin extends Outpoint {
-    value: number;
-    status: Status;
-}
-
-export interface VirtualCoin extends Coin {
-    virtualStatus: VirtualStatus;
-    spentBy?: string;
-    createdAt: Date;
-}
-
-export enum TxType {
-    TxSent = "SENT",
-    TxReceived = "RECEIVED",
-}
-
-export interface TxKey {
-    boardingTxid: string;
-    roundTxid: string;
-    redeemTxid: string;
-}
-
-export interface ArkTransaction {
-    key: TxKey;
-    type: TxType;
-    amount: number;
-    settled: boolean;
-    createdAt: number;
-}
-
-export interface IWallet {
-    // Address and balance management
-    getAddress(): Promise<AddressInfo>;
-    getBalance(): Promise<WalletBalance>;
-    getCoins(): Promise<Coin[]>;
-    getVtxos(): Promise<(SpendableVtxo & VirtualCoin)[]>;
-    getBoardingUtxos(): Promise<(SpendableVtxo & Coin)[]>;
-    getTransactionHistory(): Promise<ArkTransaction[]>;
-
-    // Transaction operations
-    sendBitcoin(params: SendBitcoinParams, zeroFee?: boolean): Promise<string>;
-    settle(
-        params?: SettleParams,
-        eventCallback?: (event: SettlementEvent) => void
-    ): Promise<string>;
-}
+} from "../../providers/ark";
+import { SignerSession } from "../signingSession";
+import { buildForfeitTx } from "../forfeit";
+import { TxWeightEstimator } from "../../utils/txSizeEstimator";
+import { validateConnectorsTree, validateVtxoTree } from "../tree/validation";
+import { Identity } from "../identity";
+import {
+    AddressInfo,
+    ArkTransaction,
+    Coin,
+    IWallet,
+    SendBitcoinParams,
+    SettleParams,
+    SpendableVtxo,
+    TxType,
+    VirtualCoin,
+    VtxoTaprootAddress,
+    WalletBalance,
+    WalletConfig,
+} from ".";
 
 // Wallet does not store any data and rely on the Ark and onchain providers to fetch utxos and vtxos
 export class Wallet implements IWallet {
