+++ test/integration.test.ts
@@ -2,8 +2,7 @@ import { expect, describe, it, beforeAll } from 'vitest'
 import { utils } from '@scure/btc-signer'
 import { hex } from '@scure/base'
 import { execSync } from 'child_process'
-import { IWallet, TxType, Wallet } from '../src/core/wallet'
-import { InMemoryKey } from '../src/core/identity'
+import { IWallet, TxType, Wallet, InMemoryKey } from '../src'
 
 const arkdExec = process.env.ARK_ENV === 'master' ? 'docker exec -t arkd' : 'nigiri'
